+++ src/hooks/useValidation.js
@@ -1,0 +1,41 @@
+import { useEffect, useState } from 'react'
+
+export const useValidation = (value, validations) => {
+  const [isEmpty, setEmpty] = useState(true)
+  const [minLengthError, setMinLengthError] = useState(false)
+  const [emailError, setEmailError] = useState(false)
+  const [inputValid, setInputValid] = useState(false)
+
+  useEffect(() => {
+    for (const validation in validations) {
+      // eslint-disable-next-line default-case
+      switch (validation) {
+        case 'minLength':
+          value.length < validations[validation] ? setMinLengthError(true) : setMinLengthError(false)
+          break
+        case 'isEmpty':
+          value ? setEmpty(false) : setEmpty(true)
+          break
+        case 'isEmail':
+          const re =
+            /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
+          re.test(String(value).toLowerCase()) ? setEmailError(false) : setEmailError(true)
+      }
+    }
+  }, [value])
+
+  useEffect(() => {
+    if (isEmpty || minLengthError || emailError) {
+      setInputValid(false)
+    } else {
+      setInputValid(true)
+    }
+  }, [isEmpty, minLengthError, emailError])
+
+  return {
+    isEmpty,
+    minLengthError,
+    emailError,
+    inputValid,
+  }
+}
