+++ src/vars/errorsValidate.js
@@ -1,0 +1,6 @@
+export const ERRORS_VALIDATE = {
+  MIN_LENGTH: 'Некорректная длина',
+  IS_EMPTY: 'Поле не должно быть пустым',
+  EMAIL: 'Некорректная почта',
+  ALL: 'Неправильно заполнены данные',
+}
